Migrate `state-viewer`'s command line interface to `clap-derive` aka `clap3-beta`. Manually tested that every command still works. Note that names of some parameter names changed: `--start-index` instead of `--start_index`.
