+++ test-utils/state-viewer/Cargo.toml
@@ -8,7 +8,8 @@ edition = "2021"
 [dependencies]
 ansi_term = "0.12"
 borsh = "0.9"
-clap = "2.33"
+clap = "=3.0.0-beta.2"
+lazy_static = "1.4"
 rayon = "1.5"
 serde = "1"
 serde_json = "1"
