+++ test-utils/state-viewer/src/lib.rs
@@ -1,7 +1,7 @@
 #![doc = include_str!("../README.md")]
 
 mod apply_chain_range;
-pub use crate::apply_chain_range::apply_chain_range;
-
 mod state_dump;
-pub use crate::state_dump::state_dump;
+
+pub mod cli;
+pub use cli::StateViewerSubCommand;
